+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /divisions/name/:name
   - /hwinfo/dtdata
   - /jobs/:id
+  - /regions/name/:region_name
   - /riak/stats
   - /stats_summary/create
   - /deliveryservices/:id/state